+++ src/core_modules/capture-core/components/FiltersForTypes/OrgUnit/types/orgUnit.types.ts
@@ -1,0 +1,10 @@
+export type OrgUnitValue = {
+    id: string;
+    name: string;
+    path: string;
+};
+
+export type OrgUnitFilterData = {
+    value: string;
+    name?: string;
+};