+++ src/core_modules/capture-core/components/FiltersForTypes/Username/usernameFilterDataGetter.ts
@@ -1,0 +1,7 @@
+import type { UsernameFilterData } from './types';
+import type { Value } from './Username.types';
+
+export const getUsernameFilterData = (value: Value): UsernameFilterData | null | undefined => {
+    if (!value) return null;
+    return { value };
+};