+++ src/core_modules/capture-core/components/ListView/Filters/FilterButton/FilterButtonMain.component.tsx
@@ -135,6 +135,7 @@ class FilterButtonMainPlain extends React.Component<Props & WithStyles<typeof ge
                 filterValue={filterValue}
                 onRemove={this.onRemove}
                 isRemovable={isRemovable}
+                onClearValue={this.handleClearFilter}
             />
         );
     }