+++ src/core_modules/capture-core/components/WorkingLists/WorkingListsCommon/helpers/buildFilterQueryArgs/filterConverters/dateTimeConverter.ts
@@ -1,0 +1,16 @@
+import type { DateTimeFilterData } from '../../../../../FiltersForTypes/DateTime';
+import { escapeString } from '../../../../../../utils/escapeString';
+
+export function convertDateTime({ sourceValue }: { sourceValue: DateTimeFilterData }): string {
+    if (sourceValue?.type !== 'ABSOLUTE') {
+        return '';
+    }
+    const requestData: string[] = [];
+    if (sourceValue.ge) {
+        requestData.push(`ge:${escapeString(sourceValue.ge)}`);
+    }
+    if (sourceValue.le) {
+        requestData.push(`le:${escapeString(sourceValue.le)}`);
+    }
+    return requestData.join(':');
+}