+++ src/core_modules/capture-core/metaDataMemoryStoreBuilders/common/factory/searchGroup/searchOperator/forcedSearchOperators.ts
@@ -19,6 +19,8 @@ export const forcedSearchOperators: Array<keyof typeof dataElementTypes> = [
     dataElementTypes.DATETIME_RANGE,
     dataElementTypes.TIME,
     dataElementTypes.TIME_RANGE,
+    dataElementTypes.PERCENTAGE,
+    dataElementTypes.PERCENTAGE_RANGE,
     dataElementTypes.BOOLEAN,
     dataElementTypes.TRUE_ONLY,
     dataElementTypes.ORGANISATION_UNIT,